+++ clang-tools-extra/clang-tidy/llvmlibc/ImplementationInNamespaceCheck.cpp
@@ -14,11 +14,16 @@ using namespace clang::ast_matchers;
 
 namespace clang::tidy::llvm_libc {
 
-const static StringRef RequiredNamespace = "__llvm_libc";
+const static StringRef RequiredNamespaceStart = "__llvm_libc";
+const static StringRef RequiredNamespaceMacroName = "LIBC_NAMESPACE";
+
 void ImplementationInNamespaceCheck::registerMatchers(MatchFinder *Finder) {
   Finder->addMatcher(
-      decl(hasParent(translationUnitDecl()), unless(linkageSpecDecl()))
-          .bind("child_of_translation_unit"),
+      translationUnitDecl(
+          forEach(decl(isExpansionInMainFile(), unless(linkageSpecDecl()),
+                       // anonymous namespaces generate usingDirective
+                       unless(usingDirectiveDecl(isImplicit())))
+                      .bind("child_of_translation_unit"))),
       this);
 }
 
@@ -26,19 +31,24 @@ void ImplementationInNamespaceCheck::check(
     const MatchFinder::MatchResult &Result) {
   const auto *MatchedDecl =
       Result.Nodes.getNodeAs<Decl>("child_of_translation_unit");
-  if (!Result.SourceManager->isInMainFile(MatchedDecl->getLocation()))
-    return;
-
-  if (const auto *NS = dyn_cast<NamespaceDecl>(MatchedDecl)) {
-    if (NS->getName() != RequiredNamespace) {
-      diag(NS->getLocation(), "'%0' needs to be the outermost namespace")
-          << RequiredNamespace;
-    }
+  MatchedDecl->dump();
+  const auto *NS = dyn_cast<NamespaceDecl>(MatchedDecl);
+  if (NS == nullptr || NS->isAnonymousNamespace()) {
+    diag(MatchedDecl->getLocation(),
+         "declaration must be enclosed within the '%0' namespace")
+        << RequiredNamespaceMacroName;
     return;
   }
-  diag(MatchedDecl->getLocation(),
-       "declaration must be declared within the '%0' namespace")
-      << RequiredNamespace;
+  if (Result.SourceManager->isMacroBodyExpansion(NS->getLocation()) == false) {
+    diag(NS->getLocation(), "the outermost namespace should be the '%0' macro")
+        << RequiredNamespaceMacroName;
+    return;
+  }
+  if (NS->getName().starts_with(RequiredNamespaceStart) == false) {
+    diag(NS->getLocation(), "the '%0' macro should start with '%1'")
+        << RequiredNamespaceMacroName << RequiredNamespaceStart;
+    return;
+  }
 }
 
 } // namespace clang::tidy::llvm_libc